redesigned into custom component
