+++ wenlambosergithubpages/package.json
@@ -23,7 +23,7 @@
     "react-device-detect": "^2.1.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.2.2",
-    "react-scripts": "^4.0.3",
+    "react-scripts": "^5.0.0",
     "typescript": "^4.6.2",
     "web-vitals": "^2.1.4",
     "web3": "^1.7.1"
